requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
build==1.2.2.post1
    # via pip-tools
cachetools==5.5.0
    # via google-auth
certifi==2024.8.30
    # via requests
charset-normalizer==3.4.0
    # via requests
click==8.1.7
    # via pip-tools
databricks-connect==16.0.0
    # via
    #   -r requirements.in
    #   databricks-dlt
databricks-dlt==0.3.0
    # via -r requirements.in
databricks-sdk==0.38.0
    # via databricks-connect
google-auth==2.36.0
    # via databricks-sdk
googleapis-common-protos==1.66.0
    # via
    #   databricks-connect
    #   grpcio-status
grpcio==1.68.1
    # via
    #   databricks-connect
    #   grpcio-status
grpcio-status==1.68.1
    # via databricks-connect
idna==3.10
    # via requests
iniconfig==2.0.0
    # via pytest
numpy==1.26.4
    # via
    #   databricks-connect
    #   pandas
packaging==24.2
    # via
    #   build
    #   databricks-connect
    #   pytest
pandas==2.2.3
    # via databricks-connect
pip-tools==7.4.1
    # via -r requirements.in
pluggy==1.5.0
    # via pytest
protobuf==5.29.0
    # via
    #   googleapis-common-protos
    #   grpcio-status
py4j==0.10.9.7
    # via databricks-connect
pyarrow==18.1.0
    # via databricks-connect
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.1
    # via google-auth
pyproject-hooks==1.2.0
    # via
    #   build
    #   pip-tools
pytest==8.3.4
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via pandas
pytz==2024.2
    # via pandas
requests==2.32.3
    # via databricks-sdk
rsa==4.9
    # via google-auth
six==1.16.0
    # via
    #   databricks-connect
    #   python-dateutil
tzdata==2024.2
    # via pandas
urllib3==2.2.3
    # via requests
wheel==0.45.1
    # via
    #   -r requirements.in
    #   pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
